HenryMaldonadoInvSystem
Hey everyone, heres the start of my inventory system. you can press tab to open a backpack panel and use your mouse to click off of it or press tab again to close it, and E is meant to pick up items off the ground but for now just deletes.